487: Applying refactoring to account.balances package
Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/487
