+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/balances/legacy/LegacyGetBalancesTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.balances
+package com.forgerock.uk.openbanking.tests.functional.account.balances.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -9,13 +9,12 @@ import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.*
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READACCOUNTSDETAIL
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READBALANCES
 
-class GetBalancesTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetBalancesTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
@@ -47,43 +46,6 @@ class GetBalancesTest(val tppResource: CreateTppCallback.TppResource) {
         // When
         val result = AccountRS().getAccountsData<OBReadBalance1>(
             accountAndTransaction3_1_2.Links.links.GetBalances, accessToken
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.balance).isNotEmpty()
-    }
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetBalances"],
-        apis = ["balances"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5", "v.3.1.4", "v.3.1.3"]
-    )
-    @Test
-    fun shouldGetBalances_v3_1_8() {
-        // Given
-        val consentRequest = OBReadConsent1().data(
-            OBReadData1()
-                .permissions(listOf(READACCOUNTSDETAIL, READBALANCES))
-        )
-            .risk(OBRisk2())
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-
-        // When
-        val result = AccountRS().getAccountsData<OBReadBalance1>(
-            accountAndTransaction3_1_8.Links.links.GetBalances, accessToken
         )
 
         // Then
